requirement.txt
flask
google-generativeai 